some naming conventions
